Move routes to dashboard app
