+++ src/apps/dashboard/routes/_asyncRoutes.ts
@@ -1,12 +1,12 @@
 import type { AsyncRoute } from 'components/router/AsyncRoute';
 
 export const ASYNC_ADMIN_ROUTES: AsyncRoute[] = [
-    { path: 'activity', page: 'dashboard/activity' },
-    { path: 'notifications', page: 'dashboard/notifications' },
-    { path: 'users/new', page: 'user/usernew' },
-    { path: 'users', page: 'user/userprofiles' },
-    { path: 'users/profile', page: 'user/useredit' },
-    { path: 'users/access', page: 'user/userlibraryaccess' },
-    { path: 'users/parentalcontrol', page: 'user/userparentalcontrol' },
-    { path: 'users/password', page: 'user/userpassword' }
+    { path: 'activity' },
+    { path: 'notifications' },
+    { path: 'users/new' },
+    { path: 'users' },
+    { path: 'users/profile' },
+    { path: 'users/access' },
+    { path: 'users/parentalcontrol' },
+    { path: 'users/password' }
 ];
